feat: move dust app endpoints under vault
