+++ front/components/app/blocks/Input.tsx
@@ -154,7 +154,7 @@ export default function Input({
                   className="ml-1 mt-2"
                   variant="secondary"
                   onClick={() => {
-                    window.location.href = `/w/${owner.sId}/a/${app.sId}/datasets/${block.config.dataset}`;
+                    window.location.href = `/w/${owner.sId}/vaults/${app.vault.sId}/apps/${app.sId}/datasets/${block.config.dataset}`;
                   }}
                   icon={PencilSquareIcon}
                   label={"Edit schema"}
