+++ front/components/vaults/VaultSideBarMenu.tsx
@@ -509,7 +509,7 @@ const VaultAppItem = ({
 }): ReactElement => {
   const router = useRouter();
 
-  const appPath = `/w/${owner.sId}/a/${app.sId}`;
+  const appPath = `/w/${owner.sId}/vaults/${app.vault.sId}/apps/${app.sId}`;
 
   return (
     <Tree.Item
