+++ front/lib/registry.ts
@@ -5,6 +5,7 @@ import config from "@app/lib/api/config";
 
 export const PRODUCTION_DUST_WORKSPACE_ID = "0ec9852c2f";
 export const PRODUCTION_DUST_APPS_WORKSPACE_ID = "78bda07b39";
+export const PRODUCTION_DUST_APPS_VAULT_ID = "vlt_rICtlrSEpWqX";
 
 export type Action = {
   app: DustAppType;
@@ -33,6 +34,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "0e9889c787",
       appHash:
         "4e896f08ef6c2c69c97610c861cd444e3d34c839eab44f9b4fd7dd1d166c40a2",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       MODEL: {
@@ -49,6 +51,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "84dfc1d4f7",
       appHash:
         "6ea231add2ae690ee959c5d8d5d06420ea2feae7dd32ac13a4e655910087e313",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       MODEL: {
@@ -64,6 +67,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "471b6aa923",
       appHash:
         "3b634a84930020a7a18d3b32f4c5f5cd85690bf4958127ba51061fb101edea33",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       DATASOURCE: {
@@ -80,6 +84,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "953b79fe89",
       appHash:
         "06e0af3c215ee205d2eff01826f763e36f5694c0650bf645ab156ee189e50b3a",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       DATASOURCE: {
@@ -102,6 +107,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "4180309c80",
       appHash:
         "8adcc9ae33a63cc735c9a23a97d7bffe658c6ef2400fc997e61e8817f611a1f8",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       SEMANTIC_SEARCH: {
@@ -125,6 +131,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "76b40f14fb",
       appHash:
         "93877e16b59a07eff3b4f154b8f568f172d6a463f27bd3bcbf5f6aa264216163",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       SUGGEST_CHANGES: {
@@ -140,6 +147,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "b4f205e453",
       appHash:
         "1ab6657b93c28ffe8a5c9678e646581cffe69bd4d6b307781a3d576da5acb03e",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       MODEL: {
@@ -155,6 +163,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "098b515f8e",
       appHash:
         "514d54c0967638656b437417228efec26de465796b5ab67ae0480d6976250768",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: { SEARCH: { provider_id: "serpapi", use_cache: false } },
   },
@@ -164,6 +173,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "21092925b9",
       appHash:
         "766618e57ff6600cac27d170395c74f4067e8671ef5bf36db5a820fb411f044b",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       WEBCONTENT: {
@@ -179,6 +189,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "d995d868a8",
       appHash:
         "7fb9c826d9de74c98de2a675093f66eab9da93a1a2cb9bc0bcc919fd074cd7eb",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       CREATE_SUGGESTIONS: {
@@ -194,6 +205,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "34a8c4a2aa",
       appHash:
         "65020161030b555f4d2efc9d1ce3a6d0020dcf76e663f746bd98213c90a0675f",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       CREATE_SUGGESTIONS: {
@@ -209,6 +221,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "b69YdlJ3PJ",
       appHash:
         "0b6b63def0224321f2bece0751bad632baca33f6d5bb596bbeb3f95b6bea5966",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       CREATE_SUGGESTIONS: {
@@ -224,6 +237,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "aba0057f4c",
       appHash:
         "e4bda2ba50f160712c08309628b4a6bf2b68dd7e9709669cc29ac43e36d663f7",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       CREATE_SUGGESTIONS: {
@@ -239,6 +253,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "b36c7416bd",
       appHash:
         "1ca7b9568681b06ef6cc0830239a479644a3ecc203c812983f3386a72e214d48",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       MODEL: {
@@ -254,6 +269,7 @@ export const DustProdActionRegistry = createActionRegistry({
       appId: "tWcuYDj1OE",
       appHash:
         "8298c6543759d1d11db0e360a8b7aa7b8ec0fa71ed274f2667678302073e4f8d",
+      appVaultId: PRODUCTION_DUST_APPS_VAULT_ID,
     },
     config: {
       MODEL: {
