+++ front/pages/w/[wId]/a.tsx
@@ -1,0 +1,36 @@
+import { withDefaultUserAuthRequirements } from "@app/lib/iam/session";
+import { VaultResource } from "@app/lib/resources/vault_resource";
+
+// This page is temporarily kept to redirect users to the global vault.
+// It should be removed once the legacy non-vault UI is removed.
+
+export const getServerSideProps = withDefaultUserAuthRequirements(
+  async (context, auth) => {
+    const owner = auth.getNonNullableWorkspace();
+    const subscription = auth.subscription();
+
+    if (!subscription) {
+      return {
+        notFound: true,
+      };
+    }
+
+    const vault = await VaultResource.fetchWorkspaceGlobalVault(auth);
+    if (!vault) {
+      return {
+        notFound: true,
+      };
+    }
+
+    return {
+      redirect: {
+        destination: !owner.flags.includes("data_vaults_feature")
+          ? `/w/${owner.sId}/vaults/${vault.sId}/apps`
+          : `/w/${owner.sId}/vaults/${vault.sId}/categories/apps`,
+        permanent: false,
+      },
+    };
+  }
+);
+
+export default function DefaultApps() {}
