+++ front/pages/w/[wId]/vaults/[vaultId]/apps/[aId]/clone.tsx
@@ -1,73 +1,86 @@
 import { Button, Tab } from "@dust-tt/sparkle";
-import type { WorkspaceType } from "@dust-tt/types";
+import type {
+  LightWorkspaceType,
+  UserTypeWithWorkspaces,
+  WorkspaceType,
+} from "@dust-tt/types";
 import type { AppType, AppVisibility } from "@dust-tt/types";
 import type { SubscriptionType } from "@dust-tt/types";
 import type { APIError } from "@dust-tt/types";
-import { ChevronRightIcon } from "@heroicons/react/24/outline";
+import { ChevronRightIcon } from "@heroicons/react/20/solid";
 import type { InferGetServerSidePropsType } from "next";
+import Link from "next/link";
 import { useRouter } from "next/router";
-import { useContext, useState } from "react";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 
-import { ConfirmContext } from "@app/components/Confirm";
 import {
   subNavigationApp,
   subNavigationBuild,
 } from "@app/components/navigation/config";
 import AppLayout from "@app/components/sparkle/AppLayout";
 import { AppLayoutSimpleCloseTitle } from "@app/components/sparkle/AppLayoutTitle";
+import WorkspacePicker from "@app/components/WorkspacePicker";
 import config from "@app/lib/api/config";
-import { withDefaultUserAuthRequirements } from "@app/lib/iam/session";
+import { getUserFromSession } from "@app/lib/iam/session";
+import { withDefaultUserAuthRequirementsNoWorkspaceCheck } from "@app/lib/iam/session";
 import { AppResource } from "@app/lib/resources/app_resource";
-import { classNames, MODELS_STRING_MAX_LENGTH } from "@app/lib/utils";
+import { classNames } from "@app/lib/utils";
 import { getDustAppsListUrl } from "@app/lib/vault_rollout";
 
-export const getServerSideProps = withDefaultUserAuthRequirements<{
-  owner: WorkspaceType;
-  subscription: SubscriptionType;
-  app: AppType;
-  dustAppsListUrl: string;
-  gaTrackingId: string;
-}>(async (context, auth) => {
-  const owner = auth.workspace();
-  const subscription = auth.subscription();
-  if (!owner || !subscription) {
-    return {
-      notFound: true,
-    };
-  }
+export const getServerSideProps =
+  withDefaultUserAuthRequirementsNoWorkspaceCheck<{
+    user: UserTypeWithWorkspaces;
+    owner: WorkspaceType;
+    subscription: SubscriptionType;
+    app: AppType;
+    dustAppsListUrl: string;
+    gaTrackingId: string;
+  }>(async (context, auth, session) => {
+    // This is a rare case where we need the full user object as we need to know the user available
+    // workspaces to clone the app into.
+    const user = await getUserFromSession(session);
+    if (!user) {
+      return {
+        notFound: true,
+      };
+    }
 
-  if (!auth.isBuilder()) {
+    const owner = auth.workspace();
+    const subscription = auth.subscription();
+
+    if (!owner || !subscription) {
+      return {
+        notFound: true,
+      };
+    }
+
+    const app = await AppResource.fetchById(
+      auth,
+      context.params?.aId as string
+    );
+
+    if (!app) {
+      return {
+        notFound: true,
+      };
+    }
+
+    const dustAppsListUrl = await getDustAppsListUrl(auth);
+
     return {
-      redirect: {
-        destination: `/w/${owner.sId}/a/${context.query.aId}`,
-        permanent: false,
+      props: {
+        user,
+        owner,
+        subscription,
+        app: app.toJSON(),
+        dustAppsListUrl,
+        gaTrackingId: config.getGaTrackingId(),
       },
     };
-  }
+  });
 
-  const app = await AppResource.fetchById(auth, context.params?.aId as string);
-
-  if (!app) {
-    return {
-      notFound: true,
-    };
-  }
-
-  const dustAppsListUrl = await getDustAppsListUrl(auth);
-
-  return {
-    props: {
-      owner,
-      subscription,
-      app: app.toJSON(),
-      dustAppsListUrl,
-      gaTrackingId: config.getGaTrackingId(),
-    },
-  };
-});
-
-export default function SettingsView({
+export default function CloneView({
+  user,
   owner,
   subscription,
   app,
@@ -78,13 +91,12 @@ export default function SettingsView({
 
   const [appName, setAppName] = useState(app.name);
   const [appNameError, setAppNameError] = useState("");
-
   const [appDescription, setAppDescription] = useState(app.description || "");
-  const [appVisibility, setAppVisibility] = useState(app.visibility);
-
-  const [isDeleting, setIsDeleting] = useState(false);
-  const [isUpdating, setIsUpdating] = useState(false);
-  const confirm = useContext(ConfirmContext);
+  const [appVisibility, setAppVisibility] = useState<AppVisibility>("private");
+  const [targetWorkspace, setTargetWorkspace] = useState<LightWorkspaceType>(
+    user.workspaces[0]
+  );
+  const [cloning, setCloning] = useState(false);
 
   const formValidation = () => {
     if (appName.length == 0) {
@@ -102,64 +114,38 @@ export default function SettingsView({
     }
   };
 
+  useEffect(() => {
+    setDisabled(!formValidation());
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [appName]);
+
   const router = useRouter();
 
-  const handleDelete = async () => {
-    if (
-      await confirm({
-        title: "Double checking",
-        message: "Are you sure you want to delete this app?",
-        validateVariant: "primaryWarning",
-      })
-    ) {
-      setIsDeleting(true);
-      const res = await fetch(`/api/w/${owner.sId}/apps/${app.sId}`, {
-        method: "DELETE",
-      });
-      if (res.ok) {
-        await router.push(dustAppsListUrl);
-      } else {
-        setIsDeleting(false);
-        const err = (await res.json()) as { error: APIError };
-        window.alert(
-          `Failed to delete the app (contact support@dust.tt for assistance) (internal error: type=${err.error.type} message=${err.error.message})`
-        );
-      }
-      return true;
-    } else {
-      return false;
-    }
-  };
-
-  const handleUpdate = async () => {
-    setIsUpdating(true);
-    const res = await fetch(`/api/w/${owner.sId}/apps/${app.sId}`, {
+  const handleClone = async () => {
+    setCloning(true);
+    const res = await fetch(`/api/w/${owner.sId}/apps/${app.sId}/clone`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
       },
       body: JSON.stringify({
-        name: appName.slice(0, MODELS_STRING_MAX_LENGTH),
-        description: appDescription.slice(0, MODELS_STRING_MAX_LENGTH),
+        name: appName,
+        description: appDescription,
         visibility: appVisibility,
+        wId: targetWorkspace.sId,
       }),
     });
     if (res.ok) {
-      await router.push(`/w/${owner.sId}/a/${app.sId}`);
-    } else {
-      setIsUpdating(false);
-      const err = (await res.json()) as { error: APIError };
-      window.alert(
-        `Failed to update the app (contact support@dust.tt for assistance) (internal error: type=${err.error.type} message=${err.error.message})`
+      const { app } = (await res.json()) as { app: AppType };
+      await router.push(
+        `/w/${targetWorkspace.sId}/vaults/${app.vault.sId}/apps/${app.sId}`
       );
+    } else {
+      const err = (await res.json()) as { error: APIError };
+      setCloning(false);
+      window.alert(`Error cloning app: ${err.error.message}`);
     }
   };
-
-  useEffect(() => {
-    setDisabled(!formValidation());
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [appName]);
 
   return (
     <AppLayout
@@ -189,7 +175,49 @@ export default function SettingsView({
           <div className="space-y-8 divide-y divide-gray-200">
             <div className="space-y-4 divide-y divide-gray-200">
               <div>
-                <div className="grid grid-cols-1 gap-x-4 gap-y-6 sm:grid-cols-6">
+                <div>
+                  <h3 className="text-base font-medium leading-6 text-gray-900">
+                    Clone <span className="ml-1 font-bold">{owner.name}</span>
+                    <ChevronRightIcon
+                      className="ml-0.5 inline h-5 w-5 pt-0.5 text-gray-500"
+                      aria-hidden="true"
+                    />
+                    <Link
+                      href={`/w/${owner.sId}/vaults/${app.vault.sId}/apps/${app.sId}`}
+                      className="w-22 mr-1 truncate text-base font-bold text-action-600 sm:w-auto"
+                    >
+                      {app.name}
+                    </Link>
+                  </h3>
+                  <p className="mt-1 text-sm text-gray-500">
+                    This will clone the app (specification along with associated
+                    datasets). You can pick a new name and description for the
+                    newly cloned app.
+                  </p>
+                </div>
+              </div>
+              <div>
+                <div className="mt-6 grid grid-cols-1 gap-x-4 gap-y-6 sm:grid-cols-6">
+                  {user.workspaces.length > 1 ? (
+                    <div className="sm:col-span-4">
+                      <label
+                        htmlFor="appName"
+                        className="block text-sm font-medium text-gray-700"
+                      >
+                        Target Workspace
+                      </label>
+                      <WorkspacePicker
+                        user={user}
+                        workspace={targetWorkspace}
+                        readOnly={false}
+                        displayDropDownOrigin="topLeft"
+                        onWorkspaceUpdate={(w) => {
+                          setTargetWorkspace(w);
+                        }}
+                      />
+                    </div>
+                  ) : null}
+
                   <div className="sm:col-span-3">
                     <label
                       htmlFor="appName"
@@ -199,7 +227,7 @@ export default function SettingsView({
                     </label>
                     <div className="mt-1 flex rounded-md shadow-sm">
                       <span className="inline-flex items-center rounded-l-md border border-r-0 border-gray-300 bg-gray-50 pl-3 pr-1 text-sm text-gray-500">
-                        {owner.name}
+                        {targetWorkspace.name}
                         <ChevronRightIcon
                           className="h-5 w-5 flex-shrink-0 pt-0.5 text-gray-400"
                           aria-hidden="true"
@@ -216,6 +244,7 @@ export default function SettingsView({
                             : "border-gray-300 focus:border-action-500 focus:ring-action-500"
                         )}
                         value={appName}
+                        readOnly={false}
                         onChange={(e) => setAppName(e.target.value)}
                       />
                     </div>
@@ -275,7 +304,7 @@ export default function SettingsView({
                             }}
                           />
                           <label
-                            htmlFor="appVisibilityPublic"
+                            htmlFor="app-visibility-public"
                             className="ml-3 block text-sm font-medium text-gray-700"
                           >
                             Public
@@ -302,7 +331,7 @@ export default function SettingsView({
                             }}
                           />
                           <label
-                            htmlFor="appVisibilityPrivate"
+                            htmlFor="app-visibility-private"
                             className="ml-3 block text-sm font-medium text-gray-700"
                           >
                             Private
@@ -313,39 +342,18 @@ export default function SettingsView({
                           </label>
                         </div>
                       </div>
-                      {appVisibility == "deleted" ? (
-                        <p className="mt-4 text-sm font-normal text-gray-500">
-                          This app is currently marked as deleted. Change its
-                          visibility above to restore it.
-                        </p>
-                      ) : null}
                     </fieldset>
                   </div>
                 </div>
               </div>
             </div>
-            <div className="flex py-6">
-              <Button
-                disabled={disable || isUpdating || isDeleting}
-                onClick={handleUpdate}
-                label={isUpdating ? "Updating..." : "Update"}
-              />
-              <div className="flex-1"></div>
+
+            <div className="pt-6">
               <div className="flex">
                 <Button
-                  variant="secondary"
-                  onClick={() => {
-                    void router.push(`/w/${owner.sId}/a/${app.sId}/clone`);
-                  }}
-                  label="Clone"
-                />
-              </div>
-              <div className="ml-2 flex">
-                <Button
-                  variant="secondaryWarning"
-                  onClick={handleDelete}
-                  disabled={isDeleting || isUpdating}
-                  label={isDeleting ? "Deleting..." : "Delete"}
+                  disabled={disable || cloning}
+                  onClick={handleClone}
+                  label={cloning ? "Cloning..." : "Clone"}
                 />
               </div>
             </div>
