+++ front/pages/w/[wId]/vaults/[vaultId]/apps/[aId]/runs/index.tsx
@@ -213,7 +213,7 @@ export default function RunsView({
                   <div className="flex items-center justify-between">
                     <div className="flex flex-initial">
                       <Link
-                        href={`/w/${owner.sId}/a/${app.sId}/runs/${run.run_id}`}
+                        href={`/w/${owner.sId}/vaults/${app.vault.sId}/apps/${app.sId}/runs/${run.run_id}`}
                         className="block"
                       >
                         <p className="font-mono truncate text-base text-action-500">
